152 params refactor (#159)
* Separated file into modular components

* Moved bitmaskselect to components

* Moved valueInput to components

* Moved rowItem to components

* Moved row to component

* Cleaned up imports

* Moved autopilot reboot modal to new file

* Moved autopilot reboot modal to new file

* Added return statement

* Update params header

* Fixed imports

* Added file description

* Added file description

* Renamed paramsSearchbar to be more descriptive

* Added comments

* Updated comments

* Fixed issue with fixedsizelist causing crash

* Fixed indent

* Added spacing

* Update function comments

* made styling more consistent by adding/removing emtpy lines and grouping tailwind colours together

---------
